requirements.txt
boto3
botocore
psycopg2-binary
pyspark==3.3.1